ref(core): Rename `Hub` to `AsyncContextStack` & remove unneeded methods
Should save some bundle size...!
